requirements.txt
Flask
pandas
matplotlib
seaborn
scikit-learn
sqlalchemy
pymysql
numpy